refactor: move everything to matrix folder
