+++ Cargo.toml
@@ -23,3 +23,4 @@ pretty_assertions = "1.3.0"
 wasm-bindgen = "0.2.84"
 tracing = "0.1.37"
 tracing-wasm = "0.2.1"
+num = "0.4.0"
