+++ src/matrix/display.rs
